Dropping attempts to run mongo with replicaset.
It's a couple weeks in and we can't connect to it.
Philipp is also researching a commercial product
and actually having *the same issues*.

The solution here might be "don't run mongo in prod".
We're going to look into what it takes to get this
project on a different backend entirely.  There's
already abstractions in place that should make it
possible.

Meanwhile, we're going back down to one mongod instance,
which actually works without a fuss, and we'll
instead make a point of ensuring our backups work.
(This may actually be easiest and most performant
to do with a GCP persistent disk snapshotting tech
than believing in mongo as well; we'll see.)

Signed-off-by: Eric Myhre <hash@exultant.us>